factuapi/requirements/base.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Task Queue
celery==5.3.4
redis==5.0.1

# File handling & PDF
Pillow==10.1.0
reportlab==4.0.7
WeasyPrint==60.1

# XML processing (for SRI integration)
lxml==4.9.3
xmlsec==1.3.13

# HTTP client
requests==2.31.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Email
django-anymail==10.2

# Monitoring & Logging
sentry-sdk==1.38.0

# Storage (for production)
django-storages==1.14.2
boto3==1.34.0